% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{climatology}
\alias{climatology}
\title{Compute climatology of grid data for a extension area of a shapefile basin.}
\usage{
climatology(
  start_date = "1981-01-01",
  end_date = "2016-12-01",
  years_omit = NULL,
  grid_path = NULL,
  name_out = NULL,
  dsn = getwd(),
  shapefile = NULL,
  main_title.plot = "Climatology",
  variable = "Precipitation\\n(mm)",
  save_GIF = TRUE,
  save_plot = TRUE,
  return_clip = TRUE
)
}
\arguments{
\item{start_date}{character. Initial date from which the calculation of the climatology will be made,
this will be introduce as 'yyyy-mm-dd'.}

\item{end_date}{character. Final date from which the calculation of the climatology will be made,
this will be introduce as 'yyyy-mm-dd'.}

\item{years_omit}{numeric. Year or Years to be omited in the climatology, this will be usefull if you data have
extreme years.}

\item{grid_path}{Character. Path where the raster stack or netcdf is stored.}

\item{name_out}{Character. Output name to save graphics files, gifs and raster stack tif format.}

\item{dsn}{chracter. Path where the output folders and files will be created.}

\item{shapefile}{SpatialPolygon. Shapefile to make the clip if \code{return_clip} is TRUE.}

\item{main_title.plot}{character. Title for plots.}

\item{variable}{character. Name of variable what you are compute climatology.}

\item{save_GIF}{logical. If TRUE returns a gif of the climatology calculate,
the output of the extension will depend on the argument \code{return_clip}. if FALSE will only create a graph for visualization.}

\item{save_plot}{logical. If TRUE returns a plot of the climatology calculate,
the output of the extension will depend on the argument \code{return_clip}. if FALSE will only create a graph for visualization.}

\item{return_clip}{logical. if this argument is TRUE and \code{shapefile} is not NULL,
all climatology outputs will be computed and returned just for shapefile extension. If FALSE
return the climatology for the original grid data.}
}
\value{
grid montlhy mean climatology of grid input.
}
\description{
Compute climatology of grid data for a extension area of a shapefile basin,
with the possibility of creating a graph of the climatology and exporting a GIF.
}
\examples{
\dontrun{
if(interactive()){
shp <- rgdal::readOGR('Data/SHPs/cuenca_mayo.shp')
t <- climatology(years_omit = c(2005,2010),
                grid_path = 'Data/GRIDs/prec_diaria.nc',
                main_title.plot = 'Mayo River basin from PISCOpd',
                shapefile = shp,
                star_date = '1981-01-01',
                end_date = '2016-12-01',
                name_out = 'mayo_basin_climatology',
                variable = 'Precipitation (mm)',
                save_GIF = TRUE,
                return_clip = TRUE,
                save_plot = FALSE)
 }
}
}
\author{
Crhistian Cornejo
}
