% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SOILGRIDS.R
\name{donwload_SOILDGRIDS}
\alias{donwload_SOILDGRIDS}
\title{donwload_SOILDGRIDS}
\usage{
donwload_SOILDGRIDS(
  path_saved = getwd(),
  variables = c("silt", "clay", "sand", "wrb"),
  layers = c("0-5cm_mean"),
  roi = c(-86, -66, -20, 2)
)
}
\arguments{
\item{path_saved}{character string to path directory where the data will save.Default: getwd()}

\item{variables}{character string of variables what you want to download. Default: c("silt", "clay", "sand", "wrb")}

\item{layers}{depth of the layers. Default: c("0-5cm_mean")}

\item{roi}{vector of Longitude min. longitude max, latitude min and latitude max. Default: c(-86, -66, -20, 2) for Peru area.}
}
\value{
download tifs of variables request.
}
\description{
Download SOILGRIDS datasets V2 for a roi area.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
