% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PERU_grids.R
\name{download_RAIN4PE}
\alias{download_RAIN4PE}
\alias{download_PISCO}
\title{Functions to download grid available data over PERU.}
\usage{
download_RAIN4PE(
  dsn = getwd(),
  openPDF = TRUE,
  shp = NULL,
  name_roi = "Cuenca_Mayo"
)

download_PISCO(
  var = c("Prec", "PET", "Temp"),
  dsn = getwd(),
  tres = c("monthly", "daily", "hourly"),
  version = "stable",
  shp = NULL,
  name_roi = NULL
)
}
\arguments{
\item{dsn}{path. Directory to which you want to store the downloaded data.}

\item{openPDF}{logical. If \code{TRUE} the PDF of the reference paper will be open. \code{FALSE} just donwload the grid data.}

\item{shp}{Path where the shapefile is alocate. Keep on \code{NULL} for download hourly data.}

\item{name_roi}{character. Name of the area or basin with which the subsetting to the grid will be done.}

\item{var}{character. variables that will be downloaded by the function, \code{c('Prec', 'PET', 'Temp')} are available.}

\item{tres}{character. Temporal resolution. \code{"daily"} or \code{"monthly"}.}

\item{version}{character. For now just 'stable' version is available.}
}
\description{
Functions to download grid available data over PERU.

Download PISCO grid daily and monthly data and subsetting for a basin extend.
}
\section{Functions}{
\itemize{
\item \code{download_RAIN4PE()}: available data for PERÚ at 0.1 x 0.1 resolution of grid and subsetting to a roi area if require.

\item \code{download_PISCO()}: RAIN4PE grid at daily scale, PISCO at daily and monthly grid data and subsetting for a roi extend.

}}
\examples{
\dontrun{
## donwloading data and subsetting for RAIN4PE
shp <- 'G:/Mi unidad/MyPackages-R/getSPPs/test/SHPs/cuenca_mayo.shp'
download_RAIN4PE(openPDF = FALSE, shp = shp, name_roi = 'Cuenca_Test')
}
\dontrun{
## donwloading data and subsetting for PISCO
var <- c('Prec', 'PET', 'Temp')
shp <- 'G:/Mi unidad/MyPackages-R/getSPPs/test/SHPs/cuenca_mayo.shp'
download_PISCO(var = var[1], dsn = getwd(), shp = shp, tres = 'monthly', version = 'stable')
}
}
\references{
Fernandez-Palomino, Carlos Antonio; Hattermann, Fred F.; Krysanova, Valentina; Lobanova, Anastasia;
Vega-Jácome, Fiorella; Lavado, Waldo; Santini, William; Aybar, Cesar; Bronstert, Axel (2021):
Rain for Peru and Ecuador (RAIN4PE). V. 1.0. GFZ Data Services. https://doi.org/10.5880/pik.2020.010

Aybar, C., Fernández, C., Huerta, A., Lavado, W., Vega, F., & Felipe-Obando, O.G. (2019).
Construction of a high-resolution gridded rainfall dataset for Peru from 1981 to the present day.
Hydrological Sciences Journal, 65, 770 - 785.
}
\seealso{
Other Download grid data: 
\code{\link{download_CMORPH}()},
\code{\link{download_SM2RAIN}()}
}
\author{
Crhistian Cornejo
}
\concept{Download grid data}
