% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CHIRPS.R
\name{download_CHIRPS}
\alias{download_CHIRPS}
\title{Function to download CHIRPS raster data global on daily and monthly scales}
\usage{
download_CHIRPS(
  tres = NULL,
  sres = NULL,
  years = 1981,
  dsn = getwd(),
  format = NULL,
  cores = parallel::detectCores() - 1,
  verbose = FALSE,
  roi = c(-86, -66, -20, 2)
)
}
\arguments{
\item{tres}{character. Temporal resolution. \code{"daily"} or \code{"monthly"}}

\item{sres}{numeric. Spatial resolution. \code{0.05} or \code{0.25}. Note
that monthly data is only available at a resolution of 0.05 degrees.}

\item{years}{numeric. Years Date(s) for which you want to download CHIRPS data.}

\item{dsn}{path. Directory to which you want to store the downloaded data.}

\item{format}{character. Download format of the files. \code{"tif"} or \code{"netcdf"}
are only available.}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}

\item{verbose}{logical. If TRUE this will show the task process for the parallel process.}

\item{roi}{numeric. Vector with xmax, xmin, ymax, ymin values for a region for subsetting the data.}
}
\description{
Function to download CHIRPS raster data global on daily and monthly scales
}
\examples{
\dontrun{
# roi = c(-86,-66,-20,2)
# download_CHIRPS(tres = 'daily',sres = 0.25,format = 'tif',years = 2000,roi = roi)
}

}
\references{
https://data.chc.ucsb.edu/products/CHIRPS-2.0/
}
\author{
Crhistian Cornejo
}
