% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMORPH.R
\name{download_CMORPH}
\alias{download_CMORPH}
\title{Function to download CMORPH raster data global on daily sacale at 25km.}
\usage{
download_CMORPH(
  star_date = "1998-01-01",
  end_date = "1998-03-01",
  dsn = getwd(),
  cores = parallel::detectCores() - 1,
  verbose = FALSE,
  subsetting = TRUE,
  roi = c(-86, -66, -20, 2)
)
}
\arguments{
\item{star_date}{numeric. Start Date(s) for which you want to download CMORPH data.}

\item{end_date}{numeric. End Date(s) for which you want to download CMORPH data. recommended to download at least one year.}

\item{dsn}{path. Directory to which you want to store the downloaded data.}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}

\item{verbose}{logical. If \code{TRUE} this will show the task process for the parallel process.}

\item{subsetting}{logical. If \code{TRUE} all downloaded netcdf's will be subsetting to the extension set to \code{roi}.}

\item{roi}{numeric. Vector with xmax, xmin, ymax, ymin values for a region for subsetting the data.}
}
\description{
Function to download CMORPH raster data global on daily sacale at 25km.
}
\examples{
\dontrun{
roi = c(-86,-66,-20,2)
download_CMORPH(star_date = '1998-01-01',
                end_date = '1998-12-31',
                roi = roi,
                subsetting = TRUE,
                verbose = FALSE)
}

}
\references{
Joyce, R. J., J. E. Janowiak, P. A. Arkin, and P. Xie, 2004:
CMORPH: A method that produces global precipitation estimates from
passive microwave and infrared data at high spatial and temporal resolution.. J. Hydromet., 5, 487-503.
}
\seealso{
Other Download grid data: 
\code{\link{donwload_SOILDGRIDS}()},
\code{\link{download_CHIRPS}()},
\code{\link{download_RAIN4PE}()},
\code{\link{download_SM2RAIN}()},
\code{\link{download_TERRACLIMATE}()}
}
\author{
Crhistian Cornejo
}
\concept{Download grid data}
