% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PERU_grids.R
\name{download_PISCO}
\alias{download_PISCO}
\title{Download PISCO grid daily and monthly data and subsetting for a basin extend.#'}
\usage{
download_PISCO(
  var = c("Prec", "PET", "Temp"),
  dsn = getwd(),
  tres = c("monthly", "daily", "hourly"),
  version = "stable",
  shp = NULL,
  name_roi = NULL
)
}
\arguments{
\item{var}{character. variables that will be downloaded by the function, \code{c('Prec', 'PET', 'Temp')} are available.}

\item{dsn}{path. Directory to which you want to store the downloaded data.}

\item{tres}{character. Temporal resolution. \code{"daily"} or \code{"monthly"}.}

\item{version}{character. For now just 'stable' version is available.}

\item{shp}{Path where the shapefile is alocate. Keep on \code{NULL} for download hourly data.}

\item{name_roi}{character. Name of the area or basin with which the subsetting to the grid will be done.}
}
\description{
Download PISCO grid daily and monthly data and subsetting for a basin extend.#'
}
\examples{
\dontrun{
## donwloading data and subsetting for PISCO
var <- c('Prec', 'PET', 'Temp')
shp <- 'G:/Mi unidad/MyPackages-R/getSPPs/test/SHPs/cuenca_mayo.shp'
download_PISCO(var = var[1], dsn = getwd(), shp = shp, tres = 'monthly', version = 'stable')
}
}
\references{
Aybar, C., Fernández, C., Huerta, A., Lavado, W., Vega, F., & Felipe-Obando, O.G. (2019).
Construction of a high-resolution gridded rainfall dataset for Peru from 1981 to the present day.
Hydrological Sciences Journal, 65, 770 - 785.
}
\author{
Crhistian Cornejo
}
