% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PERU_grids.R
\name{download_RAIN4PE}
\alias{download_RAIN4PE}
\title{Download available data for PERÚ at 0.1 x 0.1 resolution of grid and subsetting to a roi area if require.}
\usage{
download_RAIN4PE(
  dsn = getwd(),
  openPDF = TRUE,
  shp = NULL,
  name_roi = "Cuenca_Mayo"
)
}
\arguments{
\item{dsn}{path. Directory to which you want to store the downloaded data}

\item{openPDF}{logical. If \code{TRUE} the PDF of the reference paper will be open. \code{FALSE} just donwload the grid data.}

\item{name_roi}{character. Name of the area or basin with which the subsetting to the grid will be done.}
}
\description{
Download available data for PERÚ at 0.1 x 0.1 resolution of grid and subsetting to a roi area if require.
}
\examples{
\dontrun{
## donwloading data and subsetting for RAIN4PE
shp <- 'G:/Mi unidad/MyPackages-R/getSPPs/test/SHPs/cuenca_mayo.shp'
download_RAIN4PE(openPDF = FALSE, shp = shp, name_roi = 'Cuenca_Test')
}
}
\references{
Fernandez-Palomino, Carlos Antonio; Hattermann, Fred F.; Krysanova, Valentina; Lobanova, Anastasia;
Vega-Jácome, Fiorella; Lavado, Waldo; Santini, William; Aybar, Cesar; Bronstert, Axel (2021):
Rain for Peru and Ecuador (RAIN4PE). V. 1.0. GFZ Data Services. https://doi.org/10.5880/pik.2020.010
}
\seealso{
Other Download grid data: 
\code{\link{donwload_SOILDGRIDS}()},
\code{\link{download_CHIRPS}()},
\code{\link{download_CMORPH}()},
\code{\link{download_SM2RAIN}()},
\code{\link{download_TERRACLIMATE}()}
}
\author{
Crhistian Cornejo
}
\concept{Download grid data}
