% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SM2RAIN_datasets.R
\name{download_SM2RAIN}
\alias{download_SM2RAIN}
\title{Function to download SM2RAIN global netcdf datasets on daily sacale.}
\usage{
download_SM2RAIN(
  dsn = getwd(),
  data_set = c("SM2RAIN-ASCAT", "GPM+SM2RAIN", "SM2RAIN-CCI"),
  years = 2007,
  cores = parallel::detectCores() - 1
)
}
\arguments{
\item{dsn}{character. Path where the output folders and files will be created.}

\item{data_set}{character. SM2RAIN data sets available for download: c('SM2RAIN-ASCAT','GPM+SM2RAIN','SM2RAIN-CCI'). for
\code{SM2RAIN-ASCAT} data available from 2007-2021, \code{GPM+SM2RAIN} data available from 2007-2018,
\code{SM2RAIN-CCI} data available from 1998-2015,}

\item{years}{numeric. Years Date(s) for which you want to download the data.}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}
}
\description{
Function to download SM2RAIN global netcdf datasets on daily sacale.
}
\examples{
\dontrun{
download_SM2RAIN(data_set = 'SM2RAIN-ASCAT',
                years = 2007:2010)
}

}
\references{
Brocca, L., Filippucci, P., Hahn, S., Ciabatta, L., Massari, C., Camici, S., Schüller, L., Bojkov, B., and Wagner, W. (2019).
SM2RAIN-ASCAT (2007–2018): global daily satellite rainfall from ASCAT soil moisture, Earth Syst. Sci. Data Discuss.,
https://doi.org/10.5194/essd-2019-48, in review.

Brocca, L., Ciabatta, L., Massari, C., Moramarco, T., Hahn, S., Hasenauer, S., Kidd, R., Dorigo, W., Wagner, W., Levizzani, V. (2014).
Soil as a natural rain gauge: estimating global rainfall from satellite soil moisture data. Journal of Geophysical Research, 119(9),
5128-5141, doi:10.1002/2014JD021489.

Wagner, W., Hahn, S., Kidd, R., Melzer, T., Bartalis, Z., Hasenauer, S., Figa, J., de Rosnay, P., Jann, A., Schneider, S.,
Komma, J., Kubu, G., Brugger, K., Aubrecht, C., Zuger, J., Gangkofner, U., Kienberger, S., Brocca, L., Wang, Y., Bloeschl,
G., Eitzinger, J., Steinnocher, K., Zeil, P., Rubel, F. (2013). The ASCAT Soil Moisture Product: A Review of its Specifications,
Validation Results, and Emerging Applications. Meteorologische Zeitschrift, 22(1), 5-33, doi:10.1127/0941-2948/2013/0399.
}
\seealso{
Other Download grid data: 
\code{\link{donwload_SOILDGRIDS}()},
\code{\link{download_CHIRPS}()},
\code{\link{download_CMORPH}()},
\code{\link{download_RAIN4PE}()},
\code{\link{download_TERRACLIMATE}()}
}
\author{
Crhistian Cornejo
}
\concept{Download grid data}
