% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TERRACLIMATE.R
\name{download_TERRACLIMATE}
\alias{download_TERRACLIMATE}
\title{Download 14 climatic variables simultaneously from the TERRACLIMATE database at 4km resolution.}
\usage{
download_TERRACLIMATE(
  vars_id = c("aet", "def", "swe", "q", "soil", "PDSI", "pet", "ppt", "srad", "tmax",
    "tmin", "vap", "vpd", "ws"),
  years = 1958:2021,
  dsn = getwd(),
  cores = parallel::detectCores() - 1
)
}
\arguments{
\item{vars_id}{character. Name of variables what you want to download. for DEFAULT it download all
variables available.}

\item{years}{integer. Years from which you want to download the information, by DEFAULT it downloads all the
available years of TERRACLIMATE of all the variables.}

\item{dsn}{character. Path where the output folders and files will be saved.}

\item{cores}{numeric. Necessary parameter to download multiple variables in parallel.}
}
\value{
Donwload TERRACLIMATE data sets to output directory.
}
\description{
Download 14 climatic variables simultaneously from the TERRACLIMATE database at 4km resolution.
}
\examples{
\dontrun{
if(interactive()){
download_TERRACLIMATE(vars_id  = c('aet','ppt'), years = 2000:2002, cores = 4)
 }
}
}
\references{
Abatzoglou, J.T., S.Z. Dobrowski, S.A. Parks, K.C. Hegewisch, 2018, Terraclimate,
a high-resolution global dataset of monthly climate and climatic water balance from 1958-2015, Scientific Data,
}
\seealso{
Other Download grid data: 
\code{\link{donwload_SOILDGRIDS}()},
\code{\link{download_CHIRPS}()},
\code{\link{download_CMORPH}()},
\code{\link{download_RAIN4PE}()},
\code{\link{download_SM2RAIN}()}
}
\concept{Download grid data}
