% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SENAMHI_data.R
\name{get_codes}
\alias{get_codes}
\title{Function to get codes of senamhi water stations.}
\usage{
get_codes(
  stations_names = NULL,
  type = c("Met", "Hid"),
  get_LongLat = TRUE,
  plotStations = TRUE
)
}
\arguments{
\item{stations_names}{character. Names of the stations we want to download and get their code.}

\item{type}{character. Type of stations from which the codes will be returned.}

\item{get_LongLat}{logical. If \code{TRUE} Longitude and Latitude will added to the output dataframe. This needed for \code{plotStations} function.}

\item{plotStations}{logical. \code{TRUE} will plot a graph of ubications of the stations and his altitude.}
}
\value{
If \code{get_LongLat} and \code{plotStations} are TRUE a list with the table and a plot of the stations is returned,
if both are FALSE, only a two-column dataframe will be returned.
}
\description{
Function to get codes of senamhi water stations.
}
\examples{
 \dontrun{
#example for get the codes and pre-process for get_SENAMHI_data function if return is a list
names <- c('Asuncion','Augusto','llica','honda','llagaden',
           'namora','jesus','la victoria','celendin')
codes <- get_codes(stations_names = names,
                  type = 'Met',
                  get_LongLat = TRUE,
                  plotStations = TRUE)

 codes_for_function <- do.call(rbind, lapply(codes[1], data.frame, stringsAsFactors=FALSE)) \%>\%
 dplyr::arrange(StationID) \%>\%
 dplyr::select(StationID) \%>\%
 purrr::as_vector()


}

}
\references{
this functions make usefull easy wrangling download data from the package senamhiR from
source("https://gitlab.com/snippets/1793256/raw");install("senamhiR")
}
\seealso{
Other SENAMHI: 
\code{\link{get_ITPA_SENAMHI}()},
\code{\link{get_SENAMHI_data}()},
\code{\link{get_stations_ITPA_SENAMHI}()}
}
\author{
Crhistian Cornejo
}
\concept{SENAMHI}
