% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENAMHI_ITPA.R, R/get_SENAMHI_ITPA.R
\name{get_ITPA_SENAMHI}
\alias{get_ITPA_SENAMHI}
\title{Function to obtain the values of Meteorological stations from the SENAMHI Advanced Thermo-Pluviometric Information.}
\usage{
get_ITPA_SENAMHI(
  year_filter,
  station = NULL,
  export_csv = TRUE,
  fast_EDA = TRUE
)

get_ITPA_SENAMHI(
  year_filter,
  station = NULL,
  export_csv = TRUE,
  fast_EDA = TRUE
)
}
\arguments{
\item{year_filter}{character string year from where the data with information available for that particular year are granted}

\item{station}{character vector of stations names optained from \code{link{stations_ITPA_SENAMHI}}}

\item{export_csv}{logical, If \code{TRUE} it will export the table as a .csv file on yout local home working directory.}

\item{fast_EDA}{logical. Compute a fast Exploratory Data Analysis of the information.}
}
\value{
If fast_EDA is TRUE will return a list with several information form SENAMHI ITPA,
firts list: 'list[\link{1}]' is a tibble with the information of stations request, second list: 'list[\link{2}]',
is the results from fast Exploratory Data ANALYSIS from the variables in the tibble. Also it export a .csv file
to your local home working directory.

If fast_EDA is TRUE will return a list with several information form SENAMHI ITPA,
firts list: 'list[\link{1}]' is a tibble with the information of stations request, second list: 'list[\link{2}]',
is the results from fast Exploratory Data ANALYSIS from the variables in the tibble. Also it export a .csv file
to your local home working directory.
}
\description{
Obtain information from SENAMHI on Advanced Thermo-Pluviometric information from existing stations, this information is available from 2011 to the present.
You can quickly generate an Exploratory Data Analysis for the entire requested data set.

Obtain information from SENAMHI on Advanced Thermo-Pluviometric information from existing stations, this information is available from 2011 to the present.
You can quickly generate an Exploratory Data Analysis for the entire requested data set.
}
\examples{
\dontrun{
if(interactive()){
station <- c('AUGUSTO WEBERBAUER', 'NAMORA')
test <- get_ITPA_SENAMHI(year_filter = '2011',
                        station = station,
                        export_csv = TRUE,
                        fast_EDA = TRUE)

tabla_senamhi <- as_tibble(test[[1]])
missing_values <- test[[3]]
 }
}
\dontrun{
if(interactive()){
station <- c('AUGUSTO WEBERBAUER', 'NAMORA')
test <- get_ITPA_SENAMHI(year_filter = '2011',
                        station = station,
                        export_csv = TRUE,
                        fast_EDA = TRUE)

tabla_senamhi <- as_tibble(test[[1]])
missing_values <- test[[3]]
 }
}
}
