% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_RSMINERVE.R
\name{grid2RSMinerve}
\alias{grid2RSMinerve}
\title{Function extracts grid data from climate raster bricks
and make a database table for later import in RSMinerve platform.}
\usage{
grid2RSMinerve(
  grid,
  subbasins,
  var_type = "Precipitation",
  crs = 32717,
  DEM,
  dsn = getwd(),
  dates_rs = list(),
  interpolation_method = "Linear",
  cores = parallel::detectCores() - 1
)
}
\arguments{
\item{grid}{raster brick or raster stack}

\item{subbasins}{PARAM_DESCRIPTION}

\item{var_type}{Either 'Temperature' or 'Precipitation'.Default: 'Precipitation'}

\item{crs}{4 digit crs code to ensure projection consistency between raster and shapefile. Default: 32717}

\item{DEM}{Digitial Elevation Model (raster) for extracting z values.}

\item{dsn}{path directory to which you want to store the downloaded data.}

\item{dates_rs}{list with star_date and end_date of where the raster brick have information, Default: list()}

\item{interpolation_method}{type of interpolation reading for RSMinerve, Default: 'Linear'}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}
}
\description{
Function extracts grid data from climate raster bricks
and make a database table for later import in RSMinerve platform.
}
\examples{
\dontrun{
if(interactive()){
DEM <- raster::raster('usefull_data/pruebas/DEM_alos_JT.tif')
subbasins <- sf::st_read('usefull_data/pruebas/Subcuencas_v006.shp')
grid_pr <- raster::brick('usefull_data/pruebas/prec_diaria.nc')

grid2RSMinerve(grid = grid_pr,
               subbasins = subbasins,
               var_type = 'Precipitation',
               crs = 32717,DEM = DEM,
               dsn = getwd(),
               interpolation_method = 'Linear',
               dates_rs = list(Start = '1981-01-01', end = '2016-12-31'))
 }
}
}
\seealso{
Other Inputs Hidrological models: 
\code{\link{grid_pcp2SWAT}()},
\code{\link{grid_tas2SWAT}()},
\code{\link{stations2RSMinerve}()}
}
\concept{Inputs Hidrological models}
