% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_SWAT.R
\name{grid_pcp2SWAT}
\alias{grid_pcp2SWAT}
\title{Function to extract Precipitation grid data from daily and make inputs for SWAT model}
\usage{
grid_pcp2SWAT(
  pr = NULL,
  Subbasins = NULL,
  DEM = NULL,
  cores = parallel::detectCores() - 1,
  Star_Date = "1981-01-01",
  Temporal = "day",
  dsn = getwd()
)
}
\arguments{
\item{pr}{brick or stack raster/netcdf grid format of precipitation.}

\item{Subbasins}{SpatialPolygonDataFrame object of subbasins generated by SWAT model.}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}

\item{dsn}{path. Directory to which you want to store the downloaded data.}

\item{DEM.}{Raster .tif format of Digital Elevation Model which was used in the hydrological model.}

\item{Star_Date.}{character. Initial date to extract the information.}

\item{Temporal.}{character. Only \code{'day'} is aceptting for swat, but if you want to extract monthly data,
write as \code{'month'}}
}
\description{
Function to extract Precipitation grid data from daily and make inputs for SWAT model
}
\examples{
\dontrun{

DEM <- raster::raster('Data/GRIDs/DEM.tif')
pr <- 'Data/GRIDs/prec_diaria.nc'
Subbasins <- rgdal::readOGR('Data/SHPs/subs.shp')
Star_Date = '1981-01-01'
Temporal = 'day'

grid_pcp2SWAT(
  pr = pr,
  Subbasins = Subbasins,
 DEM = DEM,
  Star_Date = '1981-01-01',
  Temporal = 'day'
)

}

}
\seealso{
Other Inputs SWAT: 
\code{\link{grid_tas2SWAT}()}
}
\concept{Inputs SWAT}
