% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydClimR-package.R
\docType{package}
\name{hydClimR-package}
\alias{hydClimR}
\alias{hydClimR-package}
\title{hydClimR: Download and Pre-Process Satelital and weather stations data in R.}
\description{
Download and Pre-Process Global Satelital Climate and handling water stations data for PERU in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cCornejoR/hydClimR}
  \item \url{https://ccornejor.github.io/hydClimR/}
  \item Report bugs at \url{https://github.com/cCornejoR/hydClimR/issues}
}

}
\author{
\strong{Maintainer}: Crhistian Cornejo \email{crhistian.cornejo03@gmail.com} (\href{https://orcid.org/0000-0001-8916-2855}{ORCID})

}
\keyword{internal}
