% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MatrixCoverage.R
\name{plot_CoverageMatrix}
\alias{plot_CoverageMatrix}
\title{Plot a matrix of coverage climate data (Precipitation, Discharge or Temperature) from
a dataframe for quality control.}
\usage{
plot_CoverageMatrix(
  data_path,
  out_name,
  xlab = NULL,
  ylab = NULL,
  cex.axis,
  cex,
  lwd_line,
  color_line
)
}
\arguments{
\item{data_path}{path directory of csv dataframe with climate data, first colum
It must be one of the dates and the rest of the columns the seasons with the climatic variable that you want to plot,
these values can be daily or monthly.}

\item{out_name}{output png name for the plot.}

\item{xlab}{changes labels of the plot on x axis.}

\item{ylab}{changes labels of the plot on y axis.}

\item{cex.axis}{size of the axis labels.}

\item{cex}{size of the labels.}

\item{lwd_line}{line thickness for the plot.}
}
\description{
Plot a matrix of coverage climate data (Precipitation, Discharge or Temperature) from
a dataframe for quality control.
}
