% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_profile.R
\name{river_profile}
\alias{river_profile}
\title{create an elevation profile of a river from a Digital Elevation Model.}
\usage{
river_profile(
  path_dem,
  path_river,
  river_name,
  pt_per_km = 1,
  spanish_labels = FALSE
)
}
\arguments{
\item{path_dem}{path directory of the digital elevation model (DEM).}

\item{path_river}{path directory of the river shapefile what you want to create the river profile elevation.}

\item{river_name}{character of name of the river what your are computting profile.}

\item{pt_per_km}{Distance for extract values from DEM in km, Default: 1}

\item{spanish_labels}{logical, If \code{TRUE} this will create the plot in spanish labe, Default: FALSE}
}
\description{
create an elevation profile of a river from a Digital Elevation Model.
}
\examples{
\dontrun{
if(interactive()){
## computting river profile using the function
pt_per_km  <- 5
river_name <- 'Río Cajamarquino'
path_river <- 'test_data/pruebas/Rio_Cajamarquino.shp'
path_dem   <- 'test_data/pruebas/DEM_alos_JT.tif'

p_river <- river_profile(

 path_dem = path_dem,
 path_river = path_river,
 river_name = river_name,
 pt_per_km = pt_per_km,
 spanish_labels = TRUE

)

print(p_river)
 }
}
}
