% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydClimR-package.R
\docType{package}
\name{RWeatherTools-package}
\alias{RWeatherTools}
\alias{RWeatherTools-package}
\title{RWeatherTools: Download and Pre-Process Satelital and weather stations data in R.}
\description{
Download and Pre-Process Global Satelital Climate and handling water stations data for PERU in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cCornejoR/hydClimR}
  \item \url{https://ccornejor.github.io/hydClimR/}
  \item Report bugs at \url{https://github.com/cCornejoR/hydClimR/issues}
}

}
\author{
\strong{Maintainer}: Crhistian E. Cornejo Ruiz \email{crhistian.cornejo03@gmail.com}

}
\keyword{internal}
