% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_RSMINERVE.R
\name{stations2RSMinerve}
\alias{stations2RSMinerve}
\title{Observerd gauge data to RS MINERVE database as input for modelling.}
\usage{
stations2RSMinerve(
  path_file,
  tz = "UTC",
  units = c("C", "mm/d", "m3/s"),
  save_path = getwd(),
  interpolation_method = "Linear",
  dates_rs = list(Start = "2007-01-01", end = "2010-12-31")
)
}
\arguments{
\item{path_file}{string path to excel file to be read, see \code{link{Template_sheet_MINERVE}} to see how to
order your observed data and the function is able to generate the csv output.
Please do not change the names of the sheets.}

\item{tz}{Time zone string to be passed to base::format. Default: 'UTC'.}

\item{units}{units of your observed data, Default: c("C", "mm/d", "m3/s").}

\item{save_path}{string path where the csv output will be saved, Default: getwd().}

\item{interpolation_method}{string name of the interpolation method to use for NA's values and PET in RS Minerve, Default: 'Linear'.}

\item{dates_rs}{list with the start and end date to generate the dates that RS minerve can read, Default: list(Start = "2007-01-01", end = "2010-12-31")}
}
\value{
DataFrame ass DataBase for read in RSMinerve.
}
\description{
Converter you excel sheets data of gauge time series into data base .csv for reading
as input into RS Minerve.
}
\examples{
\dontrun{
if(interactive()){
path_file <- 'data/Stations_MINERVE.xlsx'

stations2RSMinerve(path_file = path_file)

 }
}
}
