% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SENAMHI_ITPA.R
\name{stations_ITPA_SENAMHI}
\alias{stations_ITPA_SENAMHI}
\title{Get the available stations filter by year(s) for use in the function \code{\link{get_ITPA_SENAMHI}}.}
\usage{
stations_ITPA_SENAMHI(
  names = NULL,
  years = 2011:2015,
  lonlat = FALSE,
  plot_stations = FALSE
)
}
\arguments{
\item{names}{vector. Vector names for use in the function to search matching with the information from the database.}

\item{years}{Interger year from where the data with information available for that particular year are granted.Default: 2011:2015}

\item{lonlat}{logical, if \code{TRUE} will returning a tibble with longitude and latitude, this is needed if you want
to plot it. Default: FALSE}

\item{plot_stations}{logical, \code{TRUE} will be plot on a Open Streat Map by leaflet the ubication of the stations matched. Default: FALSE}
}
\value{
Names of the stations to use in \code{\link{get_ITPA_SENAMHI}}
}
\description{
Get the available stations filter by year(s) for use in the function \code{\link{get_ITPA_SENAMHI}}.
}
\examples{
\dontrun{
if(interactive()){
names <- c('namora', 'celendin','el tigre','asuncion','papayal')

t <- stations_ITPA_SENAMHI(names = names,
                          years = 2011:2012,
                          lonlat = TRUE,
                          plot_stations = TRUE)
 }
}
}
