% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMORPH.R
\name{subsetting_CMORPH}
\alias{subsetting_CMORPH}
\title{Function to subset CMORPH raster data to a roi extension.}
\usage{
subsetting_CMORPH(
  path_files = getwd(),
  roi = c(-86, -66, -20, 2),
  vars = list(Lat = "lat", Long = "lon", main_var = "cmorph", time = "time"),
  file_out = sub("[.]nc", "_subsetting.nc", path_files),
  cores = parallel::detectCores() - 1
)
}
\arguments{
\item{path_files}{list. Path to the files where the CMORPH.nc are downloaded.}

\item{roi}{numeric. Vector with xmax, xmin, ymax, ymin values for a region for subsetting the data.1}

\item{vars}{list. Variables necessary to perform the subsetting, do not change it for CMORPH grids.}

\item{file_out}{path. Path where the new files will be saved. By default they are saved in the same path as
the downloaded files and are automatically renamed.}

\item{cores}{numeric. Number of cores to be used for the download. By default,
it automatically detects how many cores there are on your computer and
subtracts 1 to prevent it from slowing down.}
}
\value{
subsetting tif.
}
\description{
Function to subset CMORPH raster data to a roi extension.
}
\examples{
\dontrun{
if(interactive()){
roi = c(-86,-66,-20,2)
files <- base::list.files(path = paste0(dsn,'/download_CMORPH/daily/0.25deg/'),
pattern = "*.nc$",all.files=TRUE, full.names= T)
subsetting_CMORPH.nc <- function(
   path_files = files,
   roi = c(-86,-66,-20,2),
   vars = list (Lat = 'lat',
                Long = 'lon',
                main_var = 'cmorph',
                time = 'time'),
   file_out = sub("[.]nc", "_subsetting.nc", path_files),
   cores = 4)
 }
}
}
\references{
Joyce, R. J., J. E. Janowiak, P. A. Arkin, and P. Xie, 2004:
CMORPH: A method that produces global precipitation estimates from
passive microwave and infrared data at high spatial and temporal resolution.. J. Hydromet., 5, 487-503.
}
\author{
Crhistian Cornejo
}
\concept{subsetting functions}
