% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TERRACLIMATE.R
\name{subsseting_TERRACLIMATE}
\alias{subsseting_TERRACLIMATE}
\title{Function to subset and aggregate TERRACLIMATE netcdfs and exported the stack as tif.}
\usage{
subsseting_TERRACLIMATE(
  vars_id = c("aet", "def", "swe", "q", "soil", "PDSI", "pet", "ppt", "srad", "tmax",
    "tmin", "vap", "vpd", "ws"),
  dsn = getwd(),
  shapefile = NULL,
  years_sub = 1958:1960,
  factor = 0.1,
  aggr_mean_yearly = FALSE,
  export_tif_stack = TRUE,
  Rcolorbrewer_pal = "Greens"
)
}
\arguments{
\item{vars_id}{character. Name of variables what you want to do the subsetting. By DEFAULT it make the subset for
all variables donwload on the folder create by \code{link{download_TERRACLIMATE}}.}

\item{dsn}{character. Path where the output folders and files will be saved, Default: getwd()}

\item{shapefile}{SpatialPolygonDataFrame. Shapefile to make the clip, it is obligatory to introduce, Default: NULL}

\item{years_sub}{integer. Years from which you want to do the subset, by DEFAULT it does to all the
available years of TERRACLIMATE download data by \code{link{download_TERRACLIMATE}}., Default: 1958:1960}

\item{factor}{numeric. Scale factor that will be multiplied to the netcdf values, Default: 0.1}

\item{aggr_mean_yearly}{logical. TRUE if you want to return the multianual aggregate raster of the variable(s). If FALSE
return the raster brick of the subsetting, Default: FALSE.}

\item{export_tif_stack}{logical. TRUE if you want to export the raster stack of the subsetting, if FALSE it will return
just the raster brick, Default: TRUE.}

\item{Rcolorbrewer_pal}{character. Name of Rcolorbrewer palet what yout want to do the plot, Default: 'Greens'}
}
\value{
Subsettings exported raster tifs.
}
\description{
Function to subset and aggregate TERRACLIMATE netcdfs and exported the stack as tif.
}
\examples{
\dontrun{
if(interactive()){
SHP <- rgdal::readOGR('Data/SHPs/subs.shp')
rst <- subsseting_TERRACLIMATE(vars_id = 'ppt',
                              shapefile = SHP,
                              years_sub = 2000:2002,
                              factor = 0.1,
                              aggr_mean_yearly = FALSE,
                              export_tif_stack = TRUE,
                              Rcolorbrewer_pal = 'Blues')
    }
}
}
\references{
Abatzoglou, J.T., S.Z. Dobrowski, S.A. Parks, K.C. Hegewisch, 2018, Terraclimate,
a high-resolution global dataset of monthly climate and climatic water balance from 1958-2015, Scientific Data,
}
\seealso{
Other subsetting functions: 
\code{\link{subsetting_CMORPH.nc}()}
}
\author{
Crhistian Cornejo#'
}
\concept{subsetting functions}
