% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SOILGRIDS.R
\name{donwload_SOILDGRIDS}
\alias{donwload_SOILDGRIDS}
\title{Download SOILGRIDS datasets V2 for a roi area.}
\usage{
donwload_SOILDGRIDS(
  path_saved = getwd(),
  variables = c("silt", "clay", "sand", "wrb"),
  layers = c("0-5cm_mean"),
  roi = c(-86, -66, -20, 2)
)
}
\arguments{
\item{path_saved}{character string to path directory where the data will save.Default: getwd()}

\item{variables}{character string of variables what you want to download. Default: c("silt", "clay", "sand", "wrb")}

\item{layers}{depth of the layers. Default: c("0-5cm_mean")}

\item{roi}{vector of Longitude min. longitude max, latitude min and latitude max. Default: c(-86, -66, -20, 2) for Peru area.}
}
\value{
download tifs of variables request.
}
\description{
Download SOILGRIDS datasets V2 for a roi area.
}
\examples{
\dontrun{
if(interactive()){
 donwload_SOILDGRIDS(path_saved = 'C:/RPackages/hydClimR/usefull_data/SOILSGRID',
                     variables = c("silt", "clay", "sand", "wrb"),
                     layers = c('0-5cm_mean','5-15cm_mean'))
 }
}
}
\seealso{
Other Download grid data: 
\code{\link{download_CHIRPS}()},
\code{\link{download_CMORPH}()},
\code{\link{download_RAIN4PE}()},
\code{\link{download_SM2RAIN}()},
\code{\link{download_TERRACLIMATE}()}
}
\concept{Download grid data}
